Refactoring MVVM Step 1. Implementing Game Step 1.
