+++ app/src/main/AndroidManifest.xml
@@ -11,17 +11,19 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.KingdomCollector"
         tools:targetApi="31">
-
         <activity
-            android:name=".AboutUs"
+            android:name=".views.juego"
             android:exported="false" />
         <activity
-            android:name=".TermsAndConditions"
+            android:name=".views.AboutUs"
+            android:exported="false" />
+        <activity
+            android:name=".views.TermsAndConditions"
             android:exported="false"
             android:label="@string/title_activity_terms_and_conditions"
             android:theme="@style/Theme.KingdomCollector.NoActionBar" />
         <activity
-            android:name=".InitialSplashScreen"
+            android:name=".views.InitialSplashScreen"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -30,12 +32,8 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name=".views.MainActivity"
             android:exported="false" />
-        <activity android:name=".activity_tablero"
-            android:exported="false"
-            android:screenOrientation="landscape"/>
-
     </application>
 
 </manifest>
