+++ app/src/main/java/cat/udl/hyperion/appmobils/kingdomcollector/views/AboutUs.java
@@ -1,8 +1,10 @@
-package cat.udl.hyperion.appmobils.kingdomcollector;
+package cat.udl.hyperion.appmobils.kingdomcollector.views;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
+
+import cat.udl.hyperion.appmobils.kingdomcollector.R;
 
 public class AboutUs extends AppCompatActivity {
