+++ app/src/main/java/cat/udl/hyperion/appmobils/kingdomcollector/views/TermsAndConditions.java
@@ -1,4 +1,4 @@
-package cat.udl.hyperion.appmobils.kingdomcollector;
+package cat.udl.hyperion.appmobils.kingdomcollector.views;
 
 import android.os.Bundle;
