+++ app/src/main/java/cat/udl/hyperion/appmobils/kingdomcollector/views/juego.java
@@ -1,0 +1,16 @@
+package cat.udl.hyperion.appmobils.kingdomcollector.views;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+import cat.udl.hyperion.appmobils.kingdomcollector.R;
+
+public class juego extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_juego);
+    }
+}
