+++ app/src/main/res/layout/activity_about_us.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AboutUs">
+    tools:context=".views.AboutUs">
     <TextView
         android:id="@+id/textView"
         android:layout_width="wrap_content"
