+++ app/src/main/res/layout/activity_initial_splash_screen.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".InitialSplashScreen">
+    tools:context=".views.InitialSplashScreen">
 
     <pl.droidsonroids.gif.GifImageView
         android:layout_width="match_parent"
