+++ app/src/main/res/layout/activity_terms_and_conditions.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".TermsAndConditions">
+    tools:context=".views.TermsAndConditions">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/app_bar"
